move static
