+++ packages/toolkit/src/view/dashboard/lib/constant.tsx
@@ -1,0 +1,14 @@
+import { Icons } from "@instill-ai/design-system";
+
+export const options = [
+    {
+        value: "pipeline",
+        label: "Pipeline",
+        icon: <Icons.Pipeline className="h-4 w-4" />,
+    },
+    {
+        value: "model",
+        label: "Model",
+        icon: <Icons.Model className="h-4 w-4" />,
+    },
+];
